+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+- Added option for an extra namespace for better organising all ids.
 
 ## [1.3.1](https://github.com/skrepr/id-type/releases/tag/1.3.1) - 2024-09-24